ai_dev_bot_platform/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary # For PostgreSQL
redis
celery[redis]
python-telegram-bot
pydantic
pydantic-settings
httpx # For making API calls to LLMs
google-generativeai # For Gemini
# openrouter-client (if a specific client library is chosen, or use httpx)
python-dotenv # For local .env file loading
# Add other common utilities like 'requests' if needed later
# For Codebase Indexing
pgvector
alembic
pytest
pytest-mock
aider-chat
stripe
supabase
email-validator
prometheus-client
qdrant-client
cryptography